requirements.txt
bcrypt==4.2.1
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
msgspec==0.19.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
